Make 'producer' field a module constant
Spam filters blacklist the previous version of the producer field (see #366). It is only a matter of time before spammers update to a newer version of this library, with the new value of the 'producer' field. This commit makes that field a module variable, which is much cleaner to patch at runtime when needed (it was already possible before anyway, just less elegant). Moving this constant to the top of the file also enhances readability.