tests: skip cases that download artifacts when http_proxy is set (#793)
It is not currently possible to fetch files through a proxy. Until this
functionality is added, the test cases will always fail so just skip
them since they are meant to test something else.

The detection is based on the presend of the http_proxy environment
variable which should be appropriate in the vast majority of cases (I
think HTTPS_PROXY may be finer but it's probably nitpicking in
practice).